requirements.txt
flask==2.0.1
flask-cors==3.0.10
Pillow==8.3.2
pytesseract==0.3.8 